// src/app/model/Anexo.ts
    export class
    Anexo {

        pkID: pkID;
        al_razsoc: string;
        al_razsoc2: string;
        al_direcc: string;
        al_apelpat: string;
        al_apelmat: string;
        al_prinom: string;
        al_segnom: string;
        al_tipper: string;
        al_tipdocide: string;
        al_nrodocide: string;
        al_domic: string;
        al_repleg: string;
        al_carleg: string;
        al_estado: string;
        al_telf1: string;
        al_telf2: string;
        al_email: string;
        al_host: string;
        al_pais: string;
        al_nacion: string;
        al_tipdetr: string;
        al_tipperc: string;
        al_ubigeo: string;
        al_codmon: string;
        al_estcon: string;
        al_cndcon: string;
        al_numdetr: string;
        al_usrcrea: string;
        al_feccrea: string;
        al_hracrea: string;
        al_usract: string;
        al_fecact: string;
        al_hraact: string;
    }
    class pkID {
        al_tipanex: string;
        al_codanex: string;
    }